move get_surface_backend to plotting.surface._utils
