+++ nilearn/plotting/surface/_utils.py
@@ -1,11 +1,54 @@
 from warnings import warn
 
+from nilearn._utils.helpers import is_matplotlib_installed, is_plotly_installed
 from nilearn.surface import (
     PolyMesh,
     SurfaceImage,
 )
 from nilearn.surface.surface import combine_hemispheres_meshes, get_data
 from nilearn.surface.utils import check_polymesh_equal
+
+
+def get_surface_backend(engine="matplotlib"):
+    """Instantiate and return the required backend engine.
+
+    Parameters
+    ----------
+    engine: :obj:`str`, default='matplotlib'
+        Name of the required backend engine. Can be 'matplotlib' or 'plotly'.
+
+    Returns
+    -------
+    backend : :class:`~nilearn.plotting.surface._backend.BaseSurfaceBackend`
+        The backend for the specified engine.
+    """
+    if engine == "matplotlib":
+        if is_matplotlib_installed():
+            from nilearn.plotting.surface._matplotlib_backend import (
+                MatplotlibBackend,
+            )
+
+            return MatplotlibBackend()
+        else:
+            raise ImportError(
+                "Using engine='matplotlib' requires that ``matplotlib`` is "
+                "installed."
+            )
+    elif engine == "plotly":
+        if is_plotly_installed():
+            from nilearn.plotting.surface._plotly_backend import PlotlyBackend
+
+            return PlotlyBackend()
+        else:
+            raise ImportError(
+                "Using engine='plotly' requires that ``plotly`` is installed."
+            )
+    else:
+        raise ValueError(
+            f"Unknown plotting engine {engine}. "
+            "Please use either 'matplotlib' or "
+            "'plotly'."
+        )
 
 
 def _check_bg_map(bg_map, hemi):
